requirements.txt
vk-api==11.9.9
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
schedule==1.2.0